优化代码结构，将音转文独立
